requirements.txt
gradio==5.32.1
modal==1.0.2
llama-index
llama-index-llms-ollama
llama-index-llms-openai
python-dotenv
google-api-python-client==2.100.0